move farming calc constants out
